+++ shiwori_app/android/app/src/main/java/com/shiwori_app/MainActivity.java
@@ -3,7 +3,46 @@ package com.shiwori_app;
 import com.facebook.react.ReactActivity;
 import it.innove.BleManagerPackage;
 public class MainActivity extends ReactActivity {
+    private ReactRootView mReactRootView;
+    private ReactInstanceManager mReactInstanceManager;
+    private @Nullable   Callback mPermissionsCallback;
+    private @Nullable PermissionListener mPermissionListener;
+ @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        ......
+}
 
+  @Override
+    protected void onResume() {
+        super.onResume();
+
+        if (mReactInstanceManager != null) {
+            mReactInstanceManager.onHostResume(this, this);
+        }
+        if (mPermissionsCallback != null) {
+            mPermissionsCallback.invoke();
+            mPermissionsCallback = null;
+        }
+    }
+
+ @Override
+    @TargetApi(Build.VERSION_CODES.M)
+    public void requestPermissions(String[] permissions, int requestCode, PermissionListener listener) {
+        mPermissionListener = listener;
+        requestPermissions(permissions, requestCode);
+    }
+    @Override
+    public void onRequestPermissionsResult(final int requestCode, @NonNull final String[] permissions, @NonNull final int[] grantResults) {
+        mPermissionsCallback = new Callback() {
+            @Override
+            public void invoke(Object... args) {
+                if (mPermissionListener != null && mPermissionListener.onRequestPermissionsResult(requestCode, permissions, grantResults)) {
+                    mPermissionListener = null;
+                }
+            }
+        };
+    }
     /**
      * Returns the name of the main component registered from JavaScript.
      * This is used to schedule rendering of the component.
@@ -12,6 +51,8 @@ public class MainActivity extends ReactActivity {
     protected String getMainComponentName() {
         return "shiwori_app";
     }
+
+    @Override
     protected List<ReactPackage> getPackages() {
         return Arrays.<ReactPackage>asList(
             new MainReactPackage(),
