+++ shiwori_app/package.json
@@ -9,6 +9,7 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "convert-string": "^0.1.0",
     "crypto-js": "^3.1.9-1",
     "expo": "^32.0.0",
     "moment": "^2.23.0",
